Document centralised simulations; rename helpers
Add Sphinx docs: docs/reference/simulations/index.rst and class pages
for centralised, hidden_vulnerability_icml_2018 and krum_nips_2017.

Rename experiment helper files from _common.py to _run.py and update
imports in experiments. Tighten type annotations to use
torch.utils.data.Dataset and krum.primitives.attacks.Attack, and
shorten docstrings / remove redundant See Also blocks.
