+++ docs/index.rst
@@ -36,6 +36,7 @@ Reference
    reference/primitives
    reference/aggregators/index
    reference/attacks/index
+   reference/simulations/index
 
 
 License
