+++ docs/reference/simulations/classes/centralised.rst
@@ -1,0 +1,7 @@
+CentralisedSimulation
+=====================
+
+.. automodule:: krum.simulations.centralised.centralised
+   :members:
+   :undoc-members:
+   :show-inheritance:
