+++ docs/reference/simulations/classes/hidden_vulnerability_icml_2018.rst
@@ -1,0 +1,12 @@
+HiddenVulnerabilitySimulation
+==============================
+
+.. automodule:: krum.simulations.centralised.hidden_vulnerability_icml_2018
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. seealso::
+
+   For the base class, see :doc:`centralised`.
+   For the NIPS 2017 counterpart, see :doc:`krum_nips_2017`.
