+++ docs/reference/simulations/classes/krum_nips_2017.rst
@@ -1,0 +1,12 @@
+KrumSimulation
+==============
+
+.. automodule:: krum.simulations.centralised.krum_nips_2017
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. seealso::
+
+   For the base class, see :doc:`centralised`.
+   For the ICML 2018 counterpart, see :doc:`hidden_vulnerability_icml_2018`.
