+++ experiments/centralised/hidden_vulnerability_icml_2018/_run.py
@@ -1,42 +1,39 @@
-"""Shared utilities for the NIPS 2017 experiments."""
-
-from __future__ import annotations
+"""Shared helper to build and run one HiddenVulnerabilitySimulation instance."""
 
 from typing import Any
 
 import torch.nn as nn
-from torch.utils.data import Dataset as TorchDataset
+from torch.utils.data import Dataset
 
 from krum.primitives.aggregators import Aggregator
-from krum.simulations.centralised import KrumSimulation
+from krum.primitives.attacks import Attack
+from krum.simulations.centralised import HiddenVulnerabilitySimulation
 
 
 def run_one_simulation(
     *,
     label: str,
     model_cls: type[nn.Module],
-    train_set: TorchDataset[Any],
-    test_set: TorchDataset[Any],
+    train_set: Dataset[Any],
+    test_set: Dataset[Any],
     aggregator: type[Aggregator],
-    attack: Any,
+    attack: type[Attack] | None = None,
     attack_kwargs: dict[str, Any] | None = None,
     n: int,
     f: int,
     rounds: int,
     batch_size: int,
     lr: float,
+    r_eta: float,
     seed: int,
-    eval_every: int = 10,
+    eval_every: int,
+    stop_attack_at: int | None = None,
     aggregator_kwargs: dict[str, Any] | None = None,
     aggregator_f: int | None = None,
 ) -> list[tuple[int, Any]]:
-    """Build and run one :class:`KrumSimulation` instance.
-
-    Returns:
-        List of ``(round, ...)`` tuples containing evaluation results.
-    """
+    """Build and run one HiddenVulnerabilitySimulation instance."""
     print(f"\n=== {label} ===")
-    sim = KrumSimulation(
+    sim = HiddenVulnerabilitySimulation(
         model_cls=model_cls,
         train_set=train_set,
         test_set=test_set,
@@ -50,6 +47,8 @@ def run_one_simulation(
         rounds=rounds,
         batch_size=batch_size,
         lr=lr,
+        r_eta=r_eta,
+        stop_attack_at=stop_attack_at,
         seed=seed,
         eval_every=eval_every,
     )
