+++ experiments/centralised/hidden_vulnerability_icml_2018/experiment_2.py
@@ -15,7 +15,7 @@ from krum.primitives.aggregators.geomed import GeoMed
 from krum.primitives.aggregators.krum import Krum
 from krum.primitives.attacks.small_perturbation import SmallPerturbationAttack
 
-from ._common import run_one_simulation
+from ._run import run_one_simulation
 from .datasets import cifar10_dataset, mnist_dataset
 from .models import CNN, MLP
