+++ experiments/centralised/hidden_vulnerability_icml_2018/experiment_3.py
@@ -6,7 +6,7 @@ Bulyan cost vs mini-batch size. No actual attack (f=0).
 from krum.primitives.aggregators.average import Average
 from krum.primitives.aggregators.bulyan import Bulyan
 
-from ._common import run_one_simulation
+from ._run import run_one_simulation
 from .datasets import cifar10_dataset, mnist_dataset
 from .models import CNN, MLP
