+++ experiments/centralised/krum_nips_2017/experiment_1.py
@@ -7,7 +7,7 @@ from krum.primitives.aggregators.average import Average
 from krum.primitives.aggregators.krum import Krum
 from krum.primitives.attacks.gaussian import GaussianAttack
 
-from ._common import run_one_simulation
+from ._run import run_one_simulation
 from .datasets import spambase_dataset
 from .models import MLPSpambase
