+++ experiments/centralised/krum_nips_2017/experiment_3.py
@@ -8,7 +8,7 @@ from krum.primitives.aggregators.krum import Krum
 from krum.primitives.aggregators.multikrum import MultiKrum
 from krum.primitives.attacks.gaussian import GaussianAttack
 
-from ._common import run_one_simulation
+from ._run import run_one_simulation
 from .datasets import spambase_dataset
 from .models import MLPSpambase
