+++ krum/simulations/centralised/hidden_vulnerability_icml_2018.py
@@ -31,10 +31,6 @@ class HiddenVulnerabilitySimulation(CentralisedSimulation):
       weight initialization** on every layer (Section 5.1).
     - Reports three metrics per evaluation: ``(train_loss, test_accuracy,
       test_loss)``.
-
-    See Also:
-        :class:`~krum.simulations.centralised.CentralisedSimulation`
-            for the full constructor parameter list.
     """
 
     def __init__(
